Fix funnel rain collection massively overcounting charges during light precipitation
weather_sum::rain_amount accumulated arbitrary integer units (1/4/8 per turn
for very_light/light/heavy precip) but the retroactive and vehicle funnel paths
passed this value directly to funnel_charges_per_turn() which expects mm/hour.

After 15 minutes of drizzle, rain_amount would be 900 (1 unit * 900 turns) but
the correct rate is 0.5 mm/hour, causing ~1800x overcounting. A standard funnel
would collect 254 charges of water from light drizzle instead of ~0.

Fix by changing rain_amount to double and accumulating actual
precip_mm_per_hour(precip) * elapsed_turns. The two broken paths then divide
by elapsed turns to recover the average mm/hour rate before the physics calc.
The real-time fill_funnels path was already correct and is unchanged.

+++ src/weather.h
@@ -112,7 +112,7 @@ struct weather_printable {
 };
 
 struct weather_sum {
-    int rain_amount = 0;
+    double rain_amount = 0.0;
     float sunlight = 0.0f;
     float radiant_exposure = 0.0f; // J/m2
     int wind_amount = 0;
